Modify the interface package structure
