+++ src/main/java/com/classzz/czzinterfaces/controllers/NodeController.java
@@ -8,6 +8,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
@@ -19,6 +20,31 @@ public class NodeController {
 
     @Autowired
     private NodeService nodeService;
+
+    @GetMapping("getHashRate")
+    public @ResponseBody
+    JSONObject getHashRate(){
+        JSONObject result = new JSONObject();
+        result.put("state", 200);
+        result.put("version", "v1.0");
+
+        try {
+            MiningInfoDto miningInfoDto = nodeService.getMainNetWorkInfo();
+
+            JSONObject data = new JSONObject();
+            data.put("block_height", miningInfoDto.getBlocks());
+            data.put("hash_rate", miningInfoDto.getNetworkhashps());
+            data.put("testnet", miningInfoDto.getTestnet());
+            result.put("data", data);
+
+        } catch (Exception e) {
+            result.put("state", 500);
+            result.put("msg", e.getMessage());
+            LOGGER.debug(e.getMessage());
+        }
+
+        return result;
+    }
 
     @RequestMapping(value = "getdifficulty")
     public @ResponseBody
